+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 0.0.9:
 
+* Added simple boid behavior for Sarudnahk. Not able to complete the dungeon yet.
 * Added a timekeeping module to log segment timing.
 * Added backend: SeqIf/SeqWhile building blocks. Removed SeqOptional.